pyproject.toml
[project]
name = "cifo-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.4",
    "pillow>=12.2.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "pykernel>=0.1.6",
]
